ajout du crud employer

+++ controllers/listEmployeController.php
@@ -1,17 +1,21 @@
 <?php
+session_start();
 require_once '../models/userModel.php';
 require_once '../confi.php';
 
 $listUser = new user;
 $list = $listUser->listUser();
 
-if(count($_POST) > 0)
-{
 $deletUser = new user;
-$deletUser->id = $_POST['id_suppression'];
-$deletUser->deleteUser();
+if (isset($_POST['delete']) && isset($_POST['id_suppression'])) {
+    $deletUser->id = $_POST['id_suppression'];
+    $deletUser->deleteUser();
 }
-
+if (count($_GET) > 0) {
+    $reachUsers = new user;
+    $reachUsers->lastname = $_GET['search'];
+    $getReachUser = $reachUsers->seachUser();
+}
 require_once '../views/includes/header.php';
 require_once '../views/employe/listEmploye.php';
 require_once '../views/includes/footer.php';
